+++ bikeability/plugin.py
@@ -1,7 +1,7 @@
 import logging.config
 
 from climatoology.app.plugin import start_plugin
-from climatoology.utility.Naturalness import NaturalnessUtility
+from climatoology.utility.naturalness import NaturalnessUtility
 from mobility_tools.ors_settings import ORSSettings
 
 from bikeability.core.operator_worker import OperatorBikeability
@@ -13,9 +13,7 @@ log = logging.getLogger(__name__)
 def init_plugin(initialized_settings: Settings, initialized_ors_settings: ORSSettings) -> int | None:
     settings = Settings()
     naturalness_utility = NaturalnessUtility(
-        host=settings.naturalness_host,
-        port=settings.naturalness_port,
-        path=settings.naturalness_path,
+        base_url=f'http://{settings.naturalness_host}{settings.naturalness_path}',
     )
     operator = OperatorBikeability(naturalness_utility, ors_settings)